+++ src/meltano/core/plugin/singer/tap.py
@@ -11,7 +11,7 @@ from meltano.core.behavior.hookable import hook
 from meltano.core.job import JobFinder, Payload
 from meltano.core.plugin.error import PluginExecutionError, PluginLacksCapabilityError
 from meltano.core.plugin_invoker import InvokerError
-from meltano.core.setting_definition import SettingDefinition
+from meltano.core.setting_definition import SettingDefinition, SettingKind
 from meltano.core.utils import file_has_data, flatten, merge, truthy
 
 from . import PluginType, SingerPlugin
@@ -71,18 +71,21 @@ class SingerTap(SingerPlugin):
         SettingDefinition(name="_catalog"),
         SettingDefinition(name="_state"),
         SettingDefinition(name="_load_schema", value="$MELTANO_EXTRACTOR_NAMESPACE"),
-        SettingDefinition(name="_select", kind="array", value=["*.*"]),
+        SettingDefinition(name="_select", kind=SettingKind.ARRAY, value=["*.*"]),
         SettingDefinition(
             name="_metadata",
             aliases=["metadata"],
-            kind="object",
+            kind=SettingKind.OBJECT,
             value={},
             value_processor="nest_object",
         ),
         SettingDefinition(
-            name="_schema", kind="object", value={}, value_processor="nest_object"
+            name="_schema",
+            kind=SettingKind.OBJECT,
+            value={},
+            value_processor="nest_object",
         ),
-        SettingDefinition(name="_select_filter", kind="array", value=[]),
+        SettingDefinition(name="_select_filter", kind=SettingKind.ARRAY, value=[]),
     ]
 
     def exec_args(self, plugin_invoker):
@@ -364,6 +367,10 @@ class SingerTap(SingerPlugin):
             ) from err
 
     def catalog_cache_key(self, plugin_invoker):
+        # Treat non-pip plugins as editable/dev-mode plugins and do not cache.
+        if plugin_invoker.plugin.pip_url is None:
+            return None
+
         # If the extractor is installed as editable, don't cache because
         # the result of discovery could change at any time.
         if plugin_invoker.plugin.pip_url.startswith("-e"):