+++ src/meltano/core/runner/dbt.py
@@ -27,12 +27,12 @@ class DbtRunner(Runner):
     def plugin_context(self):
         return self.context.transformer
 
-    async def invoke(self, dbt: PluginInvoker, cmd, *args, log=None, **kwargs):
+    async def invoke(self, dbt: PluginInvoker, *args, log=None, **kwargs):
+        """Call the dbt executable with the given arguments in a new process."""
         log = log or sys.stderr
 
         try:
             handle = await dbt.invoke_async(
-                cmd,
                 *args,
                 **kwargs,
                 stdout=asyncio.subprocess.PIPE,
@@ -52,22 +52,17 @@ class DbtRunner(Runner):
 
         exitcode = handle.returncode
         if exitcode:
+            command = kwargs["command"] or args[0]
             raise RunnerError(
-                f"`dbt {cmd}` failed", {PluginType.TRANSFORMERS: exitcode}
+                f"`dbt {command}` failed", {PluginType.TRANSFORMERS: exitcode}
             )
 
     async def run(self, log=None):
         dbt = self.context.transformer_invoker()
 
         with dbt.prepared(self.context.session):
-            await self.invoke(dbt, "clean", log=log)
-            await self.invoke(dbt, "deps", log=log)
+            await self.invoke(dbt, log=log, command="clean")
+            await self.invoke(dbt, log=log, command="deps")
 
             cmd = "compile" if self.context.dry_run else "run"
-            await self.invoke(
-                dbt,
-                cmd,
-                "--models",
-                str(self.plugin_context.get_config("models")),
-                log=log,
-            )
+            await self.invoke(dbt, log=log, command=cmd)